Fix repeated CssImport with the same values.
Split scanner tests by type
