+++ flow-server/src/main/java/com/vaadin/flow/server/frontend/scanner/CssAnnotationVisitor.java
@@ -1,6 +1,21 @@
+/*
+ * Copyright 2000-2018 Vaadin Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
 package com.vaadin.flow.server.frontend.scanner;
 
-import java.util.HashSet;
+import java.util.List;
 
 import net.bytebuddy.jar.asm.AnnotationVisitor;
 
@@ -11,16 +26,22 @@ import com.vaadin.flow.component.dependency.CssImport;
  */
 final class CssAnnotationVisitor extends RepeatedAnnotationVisitor {
     private CssData cssData;
-    private final HashSet<CssData> cssSet;
-    
-    CssAnnotationVisitor(HashSet<CssData> cssSet) {
-        this.cssSet = cssSet;
+    private final List<CssData> cssList;
+
+    /**
+     * This visitor needs a list to be updated.
+     * 
+     * @param cssList
+     *            the list to update with this annotation values
+     */
+    CssAnnotationVisitor(List<CssData> cssList) {
+        this.cssList = cssList;
     }
 
     private void newData() {
         cssData = new CssData();
-        if (cssSet != null) {
-            cssSet.add(cssData);
+        if (cssList != null) {
+            cssList.add(cssData);
         }
     }
