+++ flow-server/src/main/java/com/vaadin/flow/server/frontend/scanner/EndPointData.java
@@ -1,9 +1,27 @@
+/*
+ * Copyright 2000-2018 Vaadin Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
 package com.vaadin.flow.server.frontend.scanner;
 
 import java.io.Serializable;
+import java.util.ArrayList;
 import java.util.Collection;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
+import java.util.stream.Collectors;
 
 import com.vaadin.flow.router.Route;
 
@@ -19,7 +37,7 @@ public final class EndPointData implements Serializable {
     ThemeData theme = new ThemeData();
     final HashSet<String> modules = new HashSet<>();
     final HashSet<String> scripts = new HashSet<>();
-    final HashSet<CssData> css = new HashSet<>();
+    final transient List<CssData> css = new ArrayList<>();
 
     private final HashSet<String> classes = new HashSet<>();
 
@@ -65,7 +83,7 @@ public final class EndPointData implements Serializable {
     }
 
     Set<CssData> getCss() {
-        return css;
+        return css.stream().collect(Collectors.toSet());
     }
 
     private String col2Str(Collection<?> s) {
